requirements.txt
langgraph
langchain
langchain-google-genai
python-dotenv
Pillow
instagrapi
torch
diffusers
transformers
accelerate
google-generativeai